Rewrote handshake code
In the old handshake, the client (the one initiating the connection)
sends the magic number and the password to the server in the clear. The
server than accepts or rejects them.

The new handshake is symmetric. Both send the magic number and a UUID
they generate, then send hash(their own UUID + the other one's UUID +
the password). They verify that the other peer sent hash(the other
peer's UUID + their own UUID + the password): this should only be the
same if they have the same password, but doesn't require sending the
password over the network on its own.

By having each peer generate a UUID, we make sure that no single peer
can decide the password's prefix. If that were possible, they could make
the prefix be "" and throw what they get back into a rainbow table to
recover the password.

We use UUIDs because it's simple to generate them and validate them.
They're free enough to have plenty of entropy, but restricted enough
that it's harder to pull off an attack requiring a chosen prefix.